Module HermesReason.Enums

val snips_result : (int, int) Stdlib.result Ctypes.typ

Contains all the enumerations related to the hermes library.

The result of a call to the hermes library.

module SNIPS_GRAIN : sig ... end

module SNIPS_HERMES_COMPONENT : sig ... end

module SNIPS_INJECTION_KIND : sig ... end

module SNIPS_PRECISION : sig ... end

module SNIPS_SESSION_INIT_TYPE : sig ... end

module SNIPS_SESSION_TERMINATION_TYPE : sig ... end

module SNIPS_SLOT_VALUE_TYPE : sig ... end