Module HermesReason.Hermes

type t = {

protocolHandler : Structs.CProtocolHandler.t_view Ctypes.ptr;

state : State.t;

}

The Hermes instance state.

val defaultOptions : Structs.CMqttOptions.t_view

Default Hermes options: no encryption and a local address on port 1883.

val make : ?⁠libraryPath:string -> ?⁠debug:bool -> ?⁠options:Structs.CMqttOptions.t_view -> unit -> t

Create a new Hermes state by connecting to the broker.

val dialogue : t -> Structs.CDialogueFacade.t_view Ctypes.ptr

val injection : t -> Structs.CInjectionFacade.t_view Ctypes.ptr

val sound : t -> Structs.CSoundFeedbackFacade.t_view Ctypes.ptr

val tts : t -> Structs.CTtsFacade.t_view Ctypes.ptr

val flowSessions : t -> string list

val perform : t -> [> `FlowEnded of string | `FlowStarted of string | `SubscribeInjectionComplete of (Structs.CInjectionCompleteMessage.t_view -> unit) * bool | `SubscribeInjectionResetComplete of (Structs.CInjectionResetCompleteMessage.t_view -> unit) * bool | `SubscribeInjectionStatus of (Structs.CInjectionStatusMessage.t_view -> unit) * bool | `SubscribeIntent of StateUtils.StringMap.key * (Structs.CIntentMessage.t_view -> unit) * bool | `SubscribeIntentNotRecognized of (Structs.CIntentNotRecognizedMessage.t_view -> unit) * bool | `SubscribeIntents of (Structs.CIntentMessage.t_view -> unit) * bool | `SubscribeSessionEnded of (Structs.CSessionEndedMessage.t_view -> unit) * bool | `SubscribeSessionQueued of (Structs.CSessionQueuedMessage.t_view -> unit) * bool | `SubscribeSessionStarted of (Structs.CSessionStartedMessage.t_view -> unit) * bool | `UnsubscribeInjectionComplete of Structs.CInjectionCompleteMessage.t_view -> unit | `UnsubscribeInjectionResetComplete of Structs.CInjectionResetCompleteMessage.t_view -> unit | `UnsubscribeInjectionStatus of Structs.CInjectionStatusMessage.t_view -> unit | `UnsubscribeIntent of StateUtils.StringMap.key * (Structs.CIntentMessage.t_view -> unit) | `UnsubscribeIntentNotRecognized of Structs.CIntentNotRecognizedMessage.t_view -> unit | `UnsubscribeIntents of Structs.CIntentMessage.t_view -> unit | `UnsubscribeSessionEnded of Structs.CSessionEndedMessage.t_view -> unit | `UnsubscribeSessionStarted of Structs.CSessionStartedMessage.t_view -> unit | `UnubscribeSessionQueued of Structs.CSessionQueuedMessage.t_view -> unit ] -> t

val destroy : t -> unit

Destroy the hermes instance and free underlying resources.