Module HermesReason__Hermes

type t = {

protocolHandler : HermesReason.Structs.CProtocolHandler.t_view Ctypes.ptr;

state : HermesReason.State.t;

}

The Hermes instance state.

val defaultOptions : HermesReason.Structs.CMqttOptions.t_view

Default Hermes options: no encryption and a local address on port 1883.

val make : ?⁠libraryPath:string -> ?⁠debug:bool -> ?⁠options:HermesReason.Structs.CMqttOptions.t_view -> unit -> t

Create a new Hermes state by connecting to the broker.

val dialogue : t -> HermesReason.Structs.CDialogueFacade.t_view Ctypes.ptr

val injection : t -> HermesReason.Structs.CInjectionFacade.t_view Ctypes.ptr

val sound : t -> HermesReason.Structs.CSoundFeedbackFacade.t_view Ctypes.ptr

val tts : t -> HermesReason.Structs.CTtsFacade.t_view Ctypes.ptr

val flowSessions : t -> string list

val perform : t -> [> `FlowEnded of string | `FlowStarted of string | `SubscribeInjectionComplete of (HermesReason.Structs.CInjectionCompleteMessage.t_view -> unit) * bool | `SubscribeInjectionResetComplete of (HermesReason.Structs.CInjectionResetCompleteMessage.t_view -> unit) * bool | `SubscribeInjectionStatus of (HermesReason.Structs.CInjectionStatusMessage.t_view -> unit) * bool | `SubscribeIntent of HermesReason.StateUtils.StringMap.key * (HermesReason.Structs.CIntentMessage.t_view -> unit) * bool | `SubscribeIntentNotRecognized of (HermesReason.Structs.CIntentNotRecognizedMessage.t_view -> unit) * bool | `SubscribeIntents of (HermesReason.Structs.CIntentMessage.t_view -> unit) * bool | `SubscribeSessionEnded of (HermesReason.Structs.CSessionEndedMessage.t_view -> unit) * bool | `SubscribeSessionQueued of (HermesReason.Structs.CSessionQueuedMessage.t_view -> unit) * bool | `SubscribeSessionStarted of (HermesReason.Structs.CSessionStartedMessage.t_view -> unit) * bool | `UnsubscribeInjectionComplete of HermesReason.Structs.CInjectionCompleteMessage.t_view -> unit | `UnsubscribeInjectionResetComplete of HermesReason.Structs.CInjectionResetCompleteMessage.t_view -> unit | `UnsubscribeInjectionStatus of HermesReason.Structs.CInjectionStatusMessage.t_view -> unit | `UnsubscribeIntent of HermesReason.StateUtils.StringMap.key * (HermesReason.Structs.CIntentMessage.t_view -> unit) | `UnsubscribeIntentNotRecognized of HermesReason.Structs.CIntentNotRecognizedMessage.t_view -> unit | `UnsubscribeIntents of HermesReason.Structs.CIntentMessage.t_view -> unit | `UnsubscribeSessionEnded of HermesReason.Structs.CSessionEndedMessage.t_view -> unit | `UnsubscribeSessionStarted of HermesReason.Structs.CSessionStartedMessage.t_view -> unit | `UnubscribeSessionQueued of HermesReason.Structs.CSessionQueuedMessage.t_view -> unit ] -> t

val destroy : t -> unit

Destroy the hermes instance and free underlying resources.